Do not delete push task names immediately
This allows task names to be tombstoned for awhile like they are
in GAE. It prevents clients from re-enqueuing the same task
multiple times.

The TaskName entities are cleaned up by the groomer after a day.
